+++ acceptance-tests/src/test/java/tech/pegasys/eth2signer/dsl/signer/Eth2SignerRunner.java
@@ -144,4 +144,8 @@ public abstract class Eth2SignerRunner {
     return signerConfig.httpPort() == UNASSIGNED_PORT;
   }
 
+  public int httpJsonRpcPort() {
+    return 0;
+  }
+
 }
